# openapi.json
{
  "components": {
    "schemas": {
...
      "AlbumItem": {
        "properties": {
          "album_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Type"
          },
          "artist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Id"
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "artist_spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Spotify Id"
          },
          "best_new": {
            "default": false,
            "title": "Best New",
            "type": "boolean"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "popularity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popularity"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_tracks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tracks"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "AlbumItem",
        "type": "object"
      },
...
      "ArtistItem": {
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "followers_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followers Count"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "popularity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popularity"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "spotify_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Url"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ArtistItem",
        "type": "object"
      },
...
      "CandidateAlbumItem": {
        "properties": {
          "album_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Type"
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "artist_spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Spotify Id"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CandidateAlbumItem",
        "type": "object"
      },
      "CandidateArtistItem": {
        "properties": {
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          }
        },
        "title": "CandidateArtistItem",
        "type": "object"
      },
      "CandidatePagination": {
        "properties": {
          "href": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Href"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "title": "CandidatePagination",
        "type": "object"
      },
      "CandidateSearchResult": {
        "properties": {
          "albums": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CandidateAlbumItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Albums"
          },
          "albums_pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidatePagination"
              },
              {
                "type": "null"
              }
            ]
          },
          "artists": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CandidateArtistItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artists"
          },
          "artists_pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidatePagination"
              },
              {
                "type": "null"
              }
            ]
          },
          "tracks": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CandidateTrackItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracks"
          },
          "tracks_pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidatePagination"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CandidateSearchResult",
        "type": "object"
      },
      "CandidateTrackAlbumRef": {
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CandidateTrackAlbumRef",
        "type": "object"
      },
      "CandidateTrackItem": {
        "properties": {
          "album": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateTrackAlbumRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "artist_spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Spotify Id"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "external_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Url"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "track_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Number"
          }
        },
        "title": "CandidateTrackItem",
        "type": "object"
      },
...
      "ExplainEntry": {
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "matched_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Field"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "popularity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popularity"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Similarity"
          }
        },
        "required": [
          "bucket",
          "id",
          "rank",
          "path"
        ],
        "title": "ExplainEntry",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrackItem": {
        "properties": {
          "album_id": {
            "title": "Album Id",
            "type": "string"
          },
          "album_spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Spotify Id"
          },
          "album_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Title"
          },
          "artist_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist Name"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "feat_artist_names": {
            "items": {
              "type": "string"
            },
            "title": "Feat Artist Names",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "spotify_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spotify Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "track_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track No"
          }
        },
        "required": [
          "id",
          "title",
          "album_id"
        ],
        "title": "TrackItem",
        "type": "object"
      },
...
      "UnifiedSearchResult": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumItem"
            },
            "title": "Albums",
            "type": "array"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistItem"
            },
            "title": "Artists",
            "type": "array"
          },
          "debug": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExplainEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackItem"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "title": "UnifiedSearchResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Music Catalog API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/music/search/candidates": {
      "get": {
        "operationId": "search_candidates_api_music_search_candidates_get",
        "parameters": [
          {
            "description": "Spotify \uac80\uc0c9 \ucffc\ub9ac",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Spotify \uac80\uc0c9 \ucffc\ub9ac",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "\ud5c8\uc6a9: \"album,artist,track\" \uc911 \uc870\ud569",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "album,artist,track",
              "description": "\ud5c8\uc6a9: \"album,artist,track\" \uc911 \uc870\ud569",
              "title": "Type",
              "type": "string"
            }
          },
          {
            "description": "\uc608: KR, US",
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\uc608: KR, US",
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "\uc120\ud0dd\uac12: \"audio\"",
            "in": "query",
            "name": "include_external",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\uc120\ud0dd\uac12: \"audio\"",
              "title": "Include External"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Spotify \ud6c4\ubcf4 \uac80\uc0c9(\uc77d\uae30 \uc804\uc6a9 + \uc568\ubc94 \ub3d9\uae30\ud654 enqueue; deprecated side effect)",
        "tags": [
          "Search"
        ]
      }
    },
    "/api/music/search/unified": {
      "get": {
        "operationId": "unified_search_api_music_search_unified_get",
        "parameters": [
          {
            "description": "\uac80\uc0c9\uc5b4",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "\uac80\uc0c9\uc5b4",
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "\uac80\uc0c9 \ub300\uc0c1 (\ucf64\ub9c8 \uc870\ud569 \ud5c8\uc6a9): \"album\", \"artist\", \"track\"",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "album,artist,track",
              "description": "\uac80\uc0c9 \ub300\uc0c1 (\ucf64\ub9c8 \uc870\ud569 \ud5c8\uc6a9): \"album\", \"artist\", \"track\"",
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Fallback offset applied to any bucket without an explicit override.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Fallback offset applied to any bucket without an explicit override.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Per-bucket offset for the artists slice (overrides `offset`).",
            "in": "query",
            "name": "artist_offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Per-bucket offset for the artists slice (overrides `offset`).",
              "title": "Artist Offset"
            }
          },
          {
            "description": "Per-bucket offset for the albums slice (overrides `offset`).",
            "in": "query",
            "name": "album_offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Per-bucket offset for the albums slice (overrides `offset`).",
              "title": "Album Offset"
            }
          },
          {
            "description": "Per-bucket offset for the tracks slice (overrides `offset`).",
            "in": "query",
            "name": "track_offset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Per-bucket offset for the tracks slice (overrides `offset`).",
              "title": "Track Offset"
            }
          },
          {
            "description": "Dev triage: include per-row ranking debug under `debug` (default response shape is otherwise unchanged).",
            "in": "query",
            "name": "explain",
            "required": false,
            "schema": {
              "default": false,
              "description": "Dev triage: include per-row ranking debug under `debug` (default response shape is otherwise unchanged).",
              "title": "Explain",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "\ud1b5\ud569 \uac80\uc0c9(DB-first)",
        "tags": [
          "Search"
        ]
      }
    },
...
  }
}